implemented Fantamaster with abstract class
