+++ packages/api/client/client.py
@@ -1,10 +1,13 @@
+from abc import ABC, abstractmethod
 import requests
+#import validators
 
-class Client:
+class Client(ABC):
 
   # TBD: lista dei moduli API disponibili e relativi endpoint
+  @abstractmethod
   def get_available_apis(self, kind=""):
-    return {}
+    raise NotImplementedError()
 
   def make_request(self, url, method="GET", headers={}, payload={}):
     try:
@@ -18,3 +21,54 @@ class Client:
       return {"data": None, "error": "Connection Error", "code": response.status_code}      
     except Exception as e:
       return {"data": None, "error":"Error while sending request: " + f"{repr(e)}", "code": None },
+
+  def __init__(self, public_api_url, tool_api_url) -> None:
+
+    if public_api_url is None or public_api_url.isspace():
+      raise ValueError("Empty Url")
+    else: self.public_api_url = public_api_url
+
+    if tool_api_url is None or tool_api_url.isspace():
+      raise ValueError("Empty Url")
+    else: self.tool_api_url = tool_api_url
+
+    #if valid: 
+    #  self.public_api_url = public_api_url
+    #else:
+    #   raise Exception("Url invalid")
+
+    #valid= validators.url(tool_api_url)
+    #if valid: 
+    #  self.tool_api_url = tool_api_url
+    #else:
+    #   raise Exception("Url invalid")
+
+  # Lista Giocatori attualmente in Serie A e utilizzabili per il fantacalcio
+  @abstractmethod
+  def playerslist(self, args={}):
+    pass
+  
+  # Giocatori attualmente indisponibili (infortunati o squalificati)
+  @abstractmethod
+  def unavailable(self, args={}):
+    pass
+
+  # Statistiche sintetiche stagione corrente
+  @abstractmethod
+  def playersstats(self, args={}):
+    pass
+
+  # Voti per ogni giornata
+  @abstractmethod
+  def ratings(self, args):
+    pass
+
+  # Probabili formazioni prossima giornata
+  @abstractmethod
+  def lineups(self, args):
+    pass
+  
+  # Dettaglio giocatore
+  @abstractmethod
+  def playerdetail(self, args):
+    pass
