--- packages/test_abc/test.py
@@ -1,4 +1,0 @@
-from fantamaster import Api
-
-api= Api()
-api.playerslist()
